requirements.txt
# Core dependencies
requests>=2.28.0
pillow>=9.0.0
tqdm>=4.64.0
psutil>=5.9.0
python-dotenv>=0.20.0
dacite>=1.6.0

# AI provider dependencies
anthropic>=0.5.0
openai>=0.27.0
ollama>=0.1.0


# NLP dependencies
nltk>=3.8.1

# Testing dependencies
pytest>=7.0.0
pytest-mock>=3.10.0
